+++ mbrc-core/Core/ApiAdapters/ILibraryApiAdapter.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using MusicBeeRemote.Core.Caching.Monitor;
 using MusicBeeRemote.Core.Model;
 using MusicBeeRemote.Core.Model.Entities;
 using MusicBeeRemote.Core.Podcasts;
@@ -7,7 +9,9 @@ namespace MusicBeeRemote.Core.ApiAdapters
 {
     public interface ILibraryApiAdapter
     {        
-        IEnumerable<Track> GetTracks();
+        IEnumerable<Track> GetTracks(string[] paths = null);
+
+        Modifications GetSyncDelta(string[] cachedFiles, DateTime lastSync);
 
         IEnumerable<Genre> GetGenres(string filter = "");
 
@@ -26,5 +30,7 @@ namespace MusicBeeRemote.Core.ApiAdapters
         IEnumerable<Playlist> GetPlaylists();
 
         bool PlayPlaylist(string url);
+        
+        IEnumerable<string> GetTrackPaths();
     }
 }