+++ mbrc-core/Core/Events/Notifications/LibraryTrackAddedEvent.cs
@@ -1,9 +1,0 @@
-﻿using TinyMessenger;
-
-namespace MusicBeeRemote.Core.Events.Notifications
-{
-    class LibraryTrackAddedEvent : ITinyMessage
-    {
-        public object Sender { get; } = null;
-    }
-}