+++ mbrc-core/Core/MusicBeeRemotePlugin.cs
@@ -104,10 +104,5 @@ namespace MusicBeeRemote.Core
         {
             _windowManager.DisplayPartyModeWindow();
         }
-
-        public void NotifyFilesAddedToLibrary()
-        {
-            _hub.Publish(new LibraryTrackAddedEvent());
-        }
     }
 }