+++ mbrc-core/Core/Settings/Dialog/Commands/RefreshLibraryCommand.cs
@@ -1,0 +1,19 @@
+﻿using MusicBeeRemote.Core.Caching.Monitor;
+
+namespace MusicBeeRemote.Core.Settings.Dialog.Commands
+{
+    public class RefreshLibraryCommand
+    {
+        private readonly ILibraryScanner _scanner;
+
+        public RefreshLibraryCommand(ILibraryScanner scanner)
+        {
+            _scanner = scanner;
+        }
+
+        public void Execute()
+        {
+            _scanner.RefreshLibrary();
+        }
+    }
+}