+++ plugin/Plugin.cs
@@ -157,9 +157,6 @@ namespace MusicBeePlugin
                 case NotificationType.NowPlayingListChanged:
                     _musicBeeRemotePlugin.NotifyNowPlayingListChanged();
                     break;
-                case NotificationType.FileAddedToLibrary:
-                    _musicBeeRemotePlugin.NotifyFilesAddedToLibrary();
-                    break;
             }
         }
     }